Move ContractClass into deprecated_contract_class
