+++ src/state_test.rs
@@ -1,4 +1,4 @@
-use crate::state::EntryPointOffset;
+use crate::deprecated_contract_class::EntryPointOffset;
 
 #[test]
 fn entry_point_offset_from_str() {
